Re-organized repo to make more sense w.r.t. bytes VS UTF-8 and complete VS streaming.
