+++ lib/ast-toolkit-snack/src/sequence/complete.rs
@@ -1,17 +1,16 @@
-//  SEQUENCE.rs
+//  COMPLETE.rs
 //    by Lut99
 //
 //  Created:
-//    04 Apr 2024, 16:24:37
+//    05 Apr 2024, 13:35:22
 //  Last edited:
-//    05 Apr 2024, 11:42:42
+//    05 Apr 2024, 13:44:49
 //  Auto updated?
 //    Yes
 //
 //  Description:
-//!   Defines combinators that apply multiple combinators in-order.
-//!   
-//!   These combinators are complete, and will directly fail if they didn't find enough input.
+//!   Defines 'complete' versions of the sequence combinators, which will
+//!   consider not enough input a hard failure.
 //
 
 use ast_toolkit_span::Span;
@@ -22,9 +21,9 @@ use crate::{Combinator, Result};
 /***** TESTS *****/
 #[cfg(test)]
 mod tests {
-    use super::super::tag;
     use super::pair;
     use crate::fail::Failure;
+    use crate::value::tag;
     use crate::Result;
 
     type Span = ast_toolkit_span::Span<&'static str, &'static str>;
