+++ lib/ast-toolkit-snack/src/sequence/mod.rs
@@ -1,0 +1,18 @@
+//  MOD.rs
+//    by Lut99
+//
+//  Created:
+//    05 Apr 2024, 13:34:35
+//  Last edited:
+//    05 Apr 2024, 13:34:46
+//  Auto updated?
+//    Yes
+//
+//  Description:
+//!   Defines combinators that parse (fixed) sequences of other
+//!   combinators.
+//
+
+// Declare submodules
+pub mod complete;
+pub mod streaming;
