+++ lib/ast-toolkit-snack/src/sequence/streaming.rs
@@ -1,0 +1,15 @@
+//  STREAMING.rs
+//    by Lut99
+//
+//  Created:
+//    05 Apr 2024, 13:35:53
+//  Last edited:
+//    05 Apr 2024, 13:36:22
+//  Auto updated?
+//    Yes
+//
+//  Description:
+//!   Defines 'complete' versions of the sequence combinators, which will
+//!   consider not enough input as a [`Failure::NotEnough`], allowing the
+//!   system to retrieve more data as necessary.
+//
