--- lib/ast-toolkit-snack/src/streaming/bytes.rs
@@ -1,14 +1,0 @@
-//  BYTES.rs
-//    by Lut99
-//
-//  Created:
-//    20 Mar 2024, 16:37:41
-//  Last edited:
-//    20 Mar 2024, 16:37:41
-//  Auto updated?
-//    Yes
-//
-//  Description:
-//!   Defines streaming combinators that are specific to operations on raw
-//!   bytes.
-//
