--- lib/ast-toolkit-snack/src/streaming/mod.rs
@@ -1,20 +1,0 @@
-//  MOD.rs
-//    by Lut99
-//
-//  Created:
-//    20 Mar 2024, 16:34:56
-//  Last edited:
-//    05 Apr 2024, 11:42:18
-//  Auto updated?
-//    Yes
-//
-//  Description:
-//!   Defines complete versions of various combinators, which assume that
-//!   more source may be on the way, and not enough input means more needs
-//!   to be requested.
-//
-
-// Declare submodules
-pub mod bytes;
-pub mod sequence;
-pub mod utf8;
