--- lib/ast-toolkit-snack/src/streaming/utf8.rs
@@ -1,14 +1,0 @@
-//  UTF 8.rs
-//    by Lut99
-//
-//  Created:
-//    20 Mar 2024, 16:37:31
-//  Last edited:
-//    20 Mar 2024, 16:37:31
-//  Auto updated?
-//    Yes
-//
-//  Description:
-//!   Defines streaming combinators that are specific to operations on
-//!   UTF-8 input.
-//
