+++ lib/ast-toolkit-snack/src/value/bytes/complete.rs
@@ -1,16 +1,16 @@
-//  BYTES.rs
+//  COMPLETE.rs
 //    by Lut99
 //
 //  Created:
-//    20 Mar 2024, 16:37:25
+//    05 Apr 2024, 13:43:32
 //  Last edited:
-//    05 Apr 2024, 13:12:15
+//    05 Apr 2024, 13:43:59
 //  Auto updated?
 //    Yes
 //
 //  Description:
-//!   Defines complete combinators that are specific to operations on raw
-//!   bytes.
+//!   Defines some raw byte-matching value combinators that are complete,
+//!   i.e., they consider not enough input a typical [`Failure`].
 //
 
 use ast_toolkit_span::{OneOfBytes, Span};
@@ -20,28 +20,6 @@ use crate::Result;
 
 
 /***** LIBRARY *****/
-/// Will attempt to match as many bytes from the start of a span as possible, as long as those bytes are in the set of to-be-searched-for bytes.
-///
-/// This version also accepts matching none of them. See [`one_of1()`] to match at least 1.
-///
-/// # Arguments
-/// - `byteset`: A byte array(-like) that defines the set of characters we are looking for.
-///
-/// # Returns
-/// A closure that will perform the actualy match for the given `byteset`. Note that this closure doesn't ever fail, because matching none is OK.
-#[inline]
-pub fn one_of0<T, F, S>(byteset: T) -> impl FnMut(Span<F, S>) -> Result<Span<F, S>, F, S>
-where
-    T: AsRef<[u8]>,
-    F: Clone,
-    S: Clone + OneOfBytes,
-{
-    move |input: Span<F, S>| -> Result<Span<F, S>, F, S> {
-        let match_point: usize = input.one_of_bytes(byteset.as_ref());
-        Result::Ok(input.slice(match_point..), input.slice(..match_point))
-    }
-}
-
 /// Will attempt to match as many bytes from the start of a span as possible, as long as those bytes are in the set of to-be-searched-for bytes.
 ///
 /// This version does _not_ accept matching none of them. See [`one_of0()`] to also allow finding none.
