+++ lib/ast-toolkit-snack/src/value/bytes/streaming.rs
@@ -1,0 +1,14 @@
+//  STREAMING.rs
+//    by Lut99
+//
+//  Created:
+//    05 Apr 2024, 13:44:11
+//  Last edited:
+//    05 Apr 2024, 13:44:23
+//  Auto updated?
+//    Yes
+//
+//  Description:
+//!   Defines some raw byte-matching value combinators that are streaming,
+//!   i.e., they consider not enough input a specific [`Failure::NotEnough`].
+//
