+++ lib/ast-toolkit-snack/src/value/utf8/streaming.rs
@@ -1,0 +1,14 @@
+//  STREAMING.rs
+//    by Lut99
+//
+//  Created:
+//    05 Apr 2024, 13:42:04
+//  Last edited:
+//    05 Apr 2024, 13:42:18
+//  Auto updated?
+//    Yes
+//
+//  Description:
+//!   Defines some UTF-8 value parsers that are streaming, i.e., they
+//!   consider not enough input as a [`Failure::NotEnough`].
+//
